Refactor and optimize target updating functions
Removed caching of game objects, making it impossible for RSR's list of game objects to not match the game's. This fixes quite a few issues such as "double summoning" of Eos/Carbuncle, delayed healing reactions, delayed mitigations, etc.
